feat: create account repository and add tests
