+++ src/features/user/repositories/user-repository/user-repository.test.ts
@@ -1,7 +1,6 @@
 import { UserRepository } from './user-repository.js';
 import { prisma } from 'mocks/prisma.js';
 import { UserMock } from '@/shared/test-helpers/mocks/user.mock.js';
-import { AccountMock } from '@/shared/test-helpers/mocks/account.mock.js';
 
 const makeSut = () => {
   const repository = new UserRepository();
@@ -77,53 +76,6 @@ describe('[Repositories] UserRepository', () => {
       });
 
       expect(result).toBeNull();
-    });
-  });
-
-  describe('getAccounts', () => {
-    it('should return user accounts if found', async () => {
-      const { repository } = makeSut();
-
-      const account = AccountMock.create();
-
-      const expectedResult = [
-        {
-          avatarUrl: account.avatarUrl,
-          createdAt: new Date(),
-          id: account.id,
-          socialMediaId: account.socialMediaId,
-          updatedAt: new Date(),
-          userId: account.userId,
-        },
-      ];
-
-      prisma.account.findMany.mockResolvedValue(expectedResult);
-
-      const result = await repository.getAccounts(account.userId);
-
-      expect(result[0]).toEqual(expectedResult[0]);
-      expect(prisma.account.findMany).toHaveBeenCalledWith({
-        where: {
-          userId: account.userId,
-        },
-      });
-    });
-
-    it('should return an empty array if user accounts are not found', async () => {
-      const { repository } = makeSut();
-
-      const userId = 'non_existent_user_id';
-
-      prisma.account.findMany.mockResolvedValue([]);
-
-      const result = await repository.getAccounts(userId);
-
-      expect(result).toEqual([]);
-      expect(prisma.account.findMany).toHaveBeenCalledWith({
-        where: {
-          userId: userId,
-        },
-      });
     });
   });
 });
